requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.5
django-countries==7.6.1
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
easy-thumbnails==2.10.1
idna==3.10
loguru==0.7.3
phonenumbers==9.0.12
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
qrcode==8.2
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
win32_setctime==1.2.0
